# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * MindBridge AI — DSA gap diagnosis tutor
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DiagnoseInput {
  /**
     * The student's free-text question or confusion statement
     * @minLength 1
     */
  question: string;
}

export interface DiagnoseResult {
  /** The topic the student appears to be asking about */
  asked_topic: string;
  /** The node id of the hypothesized missing prerequisite */
  hypothesis_node: string;
  /** LLM confidence score 0.0–1.0 */
  confidence: number;
  /** True if the LLM call failed and a fallback node was used */
  fallback: boolean;
}

export interface NodeInput {
  /** The graph node id */
  node_id: string;
}

export interface ValidationQuestion {
  node_id: string;
  label: string;
  prompt: string;
  options: string[];
  correct: string;
}

export interface TeachingContent {
  node_id: string;
  label: string;
  explanation: string;
  analogy: string;
}

export interface QuizQuestion {
  q: string;
  options: string[];
  correct: number;
}

export interface QuizResult {
  node_id: string;
  label: string;
  quiz: QuizQuestion[];
}

/**
 * Map of node_id to mastery score (0.0–1.0)
 */
export type MasteryInputMastery = {[key: string]: number};

export interface MasteryInput {
  /** Map of node_id to mastery score (0.0–1.0) */
  mastery: MasteryInputMastery;
}

export interface TopicProgress {
  id: string;
  label: string;
  mastery: number;
  prerequisites_met: boolean;
  is_unlocked: boolean;
}

export interface ProgressResult {
  /**
     * The node id of the recommended next topic, or null if all mastered
     * @nullable
     */
  next_node: string | null;
  /** @nullable */
  next_label?: string | null;
  topics: TopicProgress[];
}

export interface ErrorResponse {
  error: string;
}