# lib/api-zod/src/generated/types/diagnoseInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * MindBridge AI — DSA gap diagnosis tutor
 * OpenAPI spec version: 0.1.0
 */

export interface DiagnoseInput {
  /**
     * The student's free-text question or confusion statement
     * @minLength 1
     */
  question: string;
}