# lib/api-zod/src/generated/types/diagnoseResult.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * MindBridge AI — DSA gap diagnosis tutor
 * OpenAPI spec version: 0.1.0
 */

export interface DiagnoseResult {
  /** The topic the student appears to be asking about */
  asked_topic: string;
  /** The node id of the hypothesized missing prerequisite */
  hypothesis_node: string;
  /** LLM confidence score 0.0–1.0 */
  confidence: number;
  /** True if the LLM call failed and a fallback node was used */
  fallback: boolean;
}