# lib/api-zod/src/generated/types/masteryInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * MindBridge AI — DSA gap diagnosis tutor
 * OpenAPI spec version: 0.1.0
 */
import type { MasteryInputMastery } from './masteryInputMastery';

export interface MasteryInput {
  /** Map of node_id to mastery score (0.0–1.0) */
  mastery: MasteryInputMastery;
}