# lib/api-zod/src/generated/types/nodeInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * MindBridge AI — DSA gap diagnosis tutor
 * OpenAPI spec version: 0.1.0
 */

export interface NodeInput {
  /** The graph node id */
  node_id: string;
}