# lib/api-zod/src/generated/types/progressResult.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * MindBridge AI — DSA gap diagnosis tutor
 * OpenAPI spec version: 0.1.0
 */
import type { TopicProgress } from './topicProgress';

export interface ProgressResult {
  /**
     * The node id of the recommended next topic, or null if all mastered
     * @nullable
     */
  next_node: string | null;
  /** @nullable */
  next_label?: string | null;
  topics: TopicProgress[];
}