# lib/api-zod/src/generated/types/quizResult.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * MindBridge AI — DSA gap diagnosis tutor
 * OpenAPI spec version: 0.1.0
 */
import type { QuizQuestion } from './quizQuestion';

export interface QuizResult {
  node_id: string;
  label: string;
  quiz: QuizQuestion[];
}