# lib/api-zod/src/generated/types/teachingContent.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * MindBridge AI — DSA gap diagnosis tutor
 * OpenAPI spec version: 0.1.0
 */

export interface TeachingContent {
  node_id: string;
  label: string;
  explanation: string;
  analogy: string;
}