# lib/api-zod/src/generated/types/topicProgress.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * MindBridge AI — DSA gap diagnosis tutor
 * OpenAPI spec version: 0.1.0
 */

export interface TopicProgress {
  id: string;
  label: string;
  mastery: number;
  prerequisites_met: boolean;
  is_unlocked: boolean;
}